Refactoring to common
