+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod cpu;
 pub mod mem;
+pub mod common;
